Rename package 'output' to 'checkpoint'.
