+++ vertical-pod-autoscaler/pkg/recommender/checkpoint/checkpoint_writer_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package output
+package checkpoint
 
 import (
 	"testing"
